- restructured view class

+++ src/groundmap/grid_map.cxx
@@ -1,4 +1,4 @@
-//  $Id: grid_map.cxx,v 1.4 2003/05/19 19:00:56 grumbel Exp $
+//  $Id: grid_map.cxx,v 1.5 2003/05/19 21:46:21 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 2000 Ingo Ruhnke <grumbel@gmx.de>
@@ -62,8 +62,8 @@ GridMap::draw (View& view)
   //std::cout << "Drawing" << std::endl;
 
   /* FIXME: This code should be shared with TileMap */
-  int tile_x_offset = -(view.get_x_offset () / 40);
-  int tile_y_offset = -(view.get_y_offset () / 40);
+  int tile_x_offset = int(-(view.get_x_offset ()) / 40);
+  int tile_y_offset = int(-(view.get_y_offset ()) / 40);
   int tile_width    = tile_x_offset + (view.get_width () / 40) + 1;
   int tile_height   = tile_y_offset + (view.get_height () / 40) + 1; 
   // FIXME: one tile more to avoid artefacts, hack, hack hack...
@@ -76,7 +76,7 @@ GridMap::draw (View& view)
 	  {
 	    if (gridmap [(width * y) + x]) // Tile is not empty
 	      {
-		gridmap [(width * y) + x]->draw (view , 
+		gridmap [(width * y) + x]->draw (view, 
 						 int(x * 40) + 20, // FIXME: Map offset needs to be
 						 int(y * 40) + 20); // FIXME: configurable;
 	      }
